+++ noodles-vcf/src/reader/record/info/field/value.rs
@@ -5,7 +5,10 @@ use noodles_core as core;
 use crate::{
     header::{record::value::map::info::Type, Number},
     reader::record::MISSING,
-    record::{info::field::Value, value},
+    record::{
+        info::field::{value::Array, Value},
+        value,
+    },
 };
 
 const DELIMITER: char = ',';
@@ -90,7 +93,7 @@ fn parse_i32_array(s: &str) -> Result<Value, ParseError> {
             _ => t.parse().map(Some).map_err(ParseError::InvalidInteger),
         })
         .collect::<Result<_, _>>()
-        .map(Value::IntegerArray)
+        .map(|values| Value::Array(Array::Integer(values)))
 }
 
 fn parse_f32(s: &str) -> Result<Value, ParseError> {
@@ -106,7 +109,7 @@ fn parse_f32_array(s: &str) -> Result<Value, ParseError> {
             _ => t.parse().map(Some).map_err(ParseError::InvalidFloat),
         })
         .collect::<Result<_, _>>()
-        .map(Value::FloatArray)
+        .map(|values| Value::Array(Array::Float(values)))
 }
 
 fn parse_flag(s: &str) -> Result<Value, ParseError> {
@@ -140,7 +143,7 @@ fn parse_char_array(s: &str) -> Result<Value, ParseError> {
             _ => parse_raw_char(t).map(Some),
         })
         .collect::<Result<_, _>>()
-        .map(Value::CharacterArray)
+        .map(|values| Value::Array(Array::Character(values)))
 }
 
 fn parse_raw_string(s: &str) -> Result<String, ParseError> {
@@ -160,7 +163,7 @@ fn parse_string_array(s: &str) -> Result<Value, ParseError> {
             _ => parse_raw_string(t).map(Some),
         })
         .collect::<Result<_, _>>()
-        .map(Value::StringArray)
+        .map(|values| Value::Array(Array::String(values)))
 }
 
 #[cfg(test)]
@@ -184,11 +187,11 @@ mod tests {
 
         assert_eq!(
             parse_value(Number::Count(2), Type::Integer, "8,13"),
-            Ok(Value::IntegerArray(vec![Some(8), Some(13)])),
+            Ok(Value::Array(Array::Integer(vec![Some(8), Some(13)]))),
         );
         assert_eq!(
             parse_value(Number::Count(2), Type::Integer, "8,."),
-            Ok(Value::IntegerArray(vec![Some(8), None])),
+            Ok(Value::Array(Array::Integer(vec![Some(8), None]))),
         );
     }
 
@@ -209,11 +212,11 @@ mod tests {
 
         assert_eq!(
             parse_value(Number::Count(2), Type::Float, "0.333,0.667"),
-            Ok(Value::FloatArray(vec![Some(0.333), Some(0.667)]))
+            Ok(Value::Array(Array::Float(vec![Some(0.333), Some(0.667)])))
         );
         assert_eq!(
             parse_value(Number::Count(2), Type::Float, "0.333,."),
-            Ok(Value::FloatArray(vec![Some(0.333), None]))
+            Ok(Value::Array(Array::Float(vec![Some(0.333), None])))
         );
     }
 
@@ -255,21 +258,21 @@ mod tests {
 
         assert_eq!(
             parse_value(Number::Count(2), Type::Character, "n,d,l,s"),
-            Ok(Value::CharacterArray(vec![
+            Ok(Value::Array(Array::Character(vec![
                 Some('n'),
                 Some('d'),
                 Some('l'),
                 Some('s')
-            ]))
+            ])))
         );
         assert_eq!(
             parse_value(Number::Count(2), Type::Character, "n,d,l,."),
-            Ok(Value::CharacterArray(vec![
+            Ok(Value::Array(Array::Character(vec![
                 Some('n'),
                 Some('d'),
                 Some('l'),
                 None
-            ]))
+            ])))
         );
     }
 
@@ -294,24 +297,24 @@ mod tests {
 
         assert_eq!(
             parse_value(Number::Count(2), Type::String, "noodles,vcf"),
-            Ok(Value::StringArray(vec![
+            Ok(Value::Array(Array::String(vec![
                 Some(String::from("noodles")),
                 Some(String::from("vcf"))
-            ]))
+            ])))
         );
         assert_eq!(
             parse_value(Number::Count(2), Type::String, "noodles,."),
-            Ok(Value::StringArray(vec![
+            Ok(Value::Array(Array::String(vec![
                 Some(String::from("noodles")),
                 None
-            ]))
+            ])))
         );
         assert_eq!(
             parse_value(Number::Count(2), Type::String, "8%25,13%25"),
-            Ok(Value::StringArray(vec![
+            Ok(Value::Array(Array::String(vec![
                 Some(String::from("8%")),
                 Some(String::from("13%"))
-            ]))
+            ])))
         );
     }
 }